Refactor log trigger config
Use the config itself as a namespace instead of separate package.
